local-deployments/MVP-1.0/weather-tool/requirements.txt
fastapi
uvicorn
python-jose
requests
